move URI DBFunctions to functions namespace
